a program written to create spyware
Hello everyone, In this project, a special spyware is create in line with the information given by the user.

NOTE: THIS PROGRAM IS FOR EDUCATIONAL PURPOSES AND MUST NEVER BE USED FOR AN ILLEGAL ACTIVITY.

NOTE2: During the installation of the program, Defender may delete some loads, be careful, otherwise the program will not work.

To run the program, simply enter the file that contains the program.
Just type "python shypy.py" in the terminal.